% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedness functions.R
\name{calculate_pairwise_relatendess}
\alias{calculate_pairwise_relatendess}
\title{Calculate pairwise relatedness}
\usage{
calculate_pairwise_relatendess(
  from.id,
  to.id,
  population.data,
  r.degree = 2,
  population.network = NULL,
  fromid.ego.network = NULL,
  toid.ego.network = NULL,
  r.type = "coef.of.relationship",
  plot.ego.networks = FALSE
)
}
\description{
Calculate pairwise relatedness
}
