% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedness functions.R
\name{get__pairwise_relatednesses_matrix}
\alias{get__pairwise_relatednesses_matrix}
\title{Get pairwise relatedness between multiple individuals}
\usage{
get__pairwise_relatednesses_matrix(
  ids = "ALL",
  population.data,
  parents.to.use = "all.avaliable",
  r.degree = 2,
  kinship.network = NULL,
  r.type = "coef.of.relationship"
)
}
\description{
Get pairwise relatedness between multiple individuals
}
