% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedness functions.R
\name{get_local_relatedness}
\alias{get_local_relatedness}
\title{Get local relatedness of an individual from a kinship matrix}
\usage{
get_local_relatedness(
  id,
  kinship.matrix,
  group.members = NULL,
  local.r.to = "all"
)
}
\description{
Might not include in the final model
}
