% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedness functions.R
\name{get_relatedness}
\alias{get_relatedness}
\title{Get pairwise relatedness from an ego-pedigree}
\usage{
get_relatedness(
  from.id,
  to.id,
  fromid.ego.network,
  population.data,
  r.type = "coef.of.relationship"
)
}
\description{
Get pairwise relatedness from an ego-pedigree
}
