% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ego network functions.R
\name{indirect_relative_iterator}
\alias{indirect_relative_iterator}
\title{Find indirect relatives from a network}
\usage{
indirect_relative_iterator(network, from.id, r.degree)
}
\description{
Find indirect relatives from a network
}
