% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relatedness functions.R
\name{is_complete}
\alias{is_complete}
\title{Tests if an ego network is complete}
\usage{
is_complete(ego.network, focal, r.degree, n.lineages = 2)
}
\description{
Tests if an ego network is complete
}
