% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ego network functions.R
\name{get_nth_degree_ego_network}
\alias{get_nth_degree_ego_network}
\title{Create ego-pedigree}
\usage{
get_nth_degree_ego_network(
  network,
  ego.id,
  r.degree = 2,
  seperate.lineages = FALSE
)
}
\description{
Create ego-pedigree
}
