% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ego network functions.R
\name{get_nth_degree_relatives}
\alias{get_nth_degree_relatives}
\title{Find relatives from a pedigree-network}
\usage{
get_nth_degree_relatives(network, from.id, r.degree)
}
\description{
Find relatives from a pedigree-network
}
