% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ego network functions.R
\name{make_relation_df}
\alias{make_relation_df}
\title{Simplify data frame to pedigree information}
\usage{
make_relation_df(population.data, parents.to.use = "all.avaliable")
}
\description{
Simplify data frame to pedigree information
}
